+++ TestEnvironment.psm1
@@ -90,6 +90,7 @@ Export-ModuleMember -Function @(
     'Remove-TestEnvironment',
     'Get-TestEnvironmentReport',
     'Test-TestEnvironment',
+    'Compare-TestEnvironment',
     'Get-TestAccessToken',
     'New-TestServiceApp',
     'Get-TestServiceApp',